1、支持细颗粒度菜单权限控制，通过勾选的方式配置erupt菜单的增删改查等按钮的使用权限 1、EruptRole移除PowerOff配置 2、增加erupt菜单时会自动生成增、删、改、查按钮菜单项

+++ erupt-i18n/src/main/java/xyz/erupt/i18n/core/I18nProcess.java
@@ -57,7 +57,7 @@ public class I18nProcess extends HashMap<String, Properties> implements Applicat
             }
         }
         i18NTranslateService.registerI18NMapping(langMappings);
-        log.info("Erupt i18n total file size: {}kb", totalSize / 1000);
+        log.info("Erupt i18n total file size: {}kb", totalSize.doubleValue() / 1000);
     }
 
     @SneakyThrows
